Measure And Generate Custom Axes

setup_y_axes(): Uses the currently active viewport to set up the y axes. The primary axes is drawn on the left, the secondary axes on the right side of the currently active viewport.

setup_x_axes(): Uses the currently active viewport to set up the x axes. The height at which it is drawn is determined by the primary y axes 0 point. If y0 is not on the y axes then the x axes will either be drawn at the bottom, if only positive values are present, or at the top, if only negative values are present.

setup_xy_axes(): Uses the currently active viewport to set up both the x and y axes.

setup_guiding_lines(): Uses the currently active viewport to set up the x and y axes guiding lines. Guiding lines are drawn at the tick positions and reach from the axes to the opposite end of the diagram.

swap_scaling(): Width and height are measured according to the individual scaling for each dimension. This function converts the measured to the respective other dimension.

even_prettier(): Generates a sequence of pretty steps for equal length ticks. The function calculates more granular results than the base pretty() function.

Usage

setup_y_axes(diagram_info, arguments, secondary = FALSE)

setup_x_axes(diagram_info, arguments)

setup_xy_axes(diagram_info, arguments)

setup_guiding_lines(diagram_info, arguments)

swap_scaling(measuring, from = "width", to = "height")

even_prettier(min_value, max_value, number_of_steps)

Arguments

diagram_info

The list of measurements generated by get_diagram_dimensions().

arguments

Argument list passed on by design_graphic().

secondary

FALSE by default. If TRUE draws the primary axes on the left side, otherwise the secondary axes on the right side.

measuring

The values received from get_text_width() or get_text_height().

from

Input measuring.

to

Output measuring.

min_value

Minimum starting value.

max_value

Maximum end value.

number_of_steps

The number of equal length steps to be used to head from the minimum to the maximum value.

Value

setup_y_axes(), setup_x_axes(), setup_xy_axes(): Returns a grid::gTree object containing the full axes.

swap_scaling(): Returns a vector of width or heights.

even_prettier(): Returns a pretty sequence of equal length numeric values.