Generate Graphics While Fully Concentrating On Design

design_graphic() produces fully styled diagrams. The many parameters are all focused around making visual changes simple and fast. All diagrams can be exported into multiple formats.When the gridSVG package is installed, it is also possible to export interactive diagrams.

The diagram parameter takes in a function for the main diagram area, making the function modular and extendable.

Usage

design_graphic(
  data_frame,
  axes_variables = "",
  segments,
  values,
  statistics = "pct_group",
  pct_value = list(),
  compute = list(),
  formats = list(),
  by = c(),
  weight = NULL,
  titles = .qol_options[["titles"]],
  footnotes = .qol_options[["footnotes"]],
  var_labels = .qol_options[["var_labels"]],
  stat_labels = .qol_options[["stat_labels"]],
  diagram = dg_vbars,
  visuals = .qol_options[["graphic_visuals"]],
  axes = .qol_options[["graphic_axes"]],
  dimensions = .qol_options[["graphic_dimensions"]],
  fine_tuning = .qol_options[["graphic_fine_tuning"]],
  add_texts = NULL,
  add_forms = NULL,
  output = .qol_options[["graphic_output"]],
  na.rm = .qol_options[["na.rm"]],
  print = .qol_options[["print"]],
  monitor = .qol_options[["monitor"]]
)

Arguments

data_frame

A data frame in which are the variables to tabulate.

axes_variables

A single variable combination or a vector with single variable that should appear on the variable axes. To nest variables use the form: "var1 + var2 + var3 + ...".

segments

A vector that provides variables that should appear as the graphics segments.

values

A vector containing all variables that should be summarised.

statistics

Available functions:

"sum" -> Weighted and unweighted sum

"sum_wgt" -> Sum of all weights

"freq" -> Unweighted frequency

"freq_g0" -> Unweighted frequency of all values greater than zero

"pct_group" -> Weighted and unweighted percentages within the respective group

"pct_value" -> Weighted and unweighted percentages between value variables

"pct_total" -> Weighted and unweighted percentages compared to the grand total

"mean" -> Weighted and unweighted mean

"median" -> Weighted and unweighted median

"mode" -> Weighted and unweighted mode

"min" -> Minimum

"max" -> Maximum

"sd" -> Weighted and unweighted standard deviation

"variance" -> Weighted and unweighted standard variance

"first" -> First value

"last" -> Last value

"pn" -> Weighted and unweighted percentiles (any p1, p2, p3, ... possible) Uses the SAS default way of calculating percentiles, which is Type 2 described in quantile().

"missing" -> Missings generated by the value variables

pct_value

If pct_value is specified in the statistics, you can pass a list here which contains the information for a new variable name and between which of the value variables percentages should be computed.

compute

A list in which individual calculations are handled. For the writing style see the compute.() function.

formats

A list in which is specified which formats should be applied to which variables.

by

Compute tables stratified by the expressions of the provided variables.

weight

Put in a weight variable to compute weighted results.

titles

Specify one or more table titles.

footnotes

Specify one or more table footnotes.

var_labels

A list in which is specified which label should be printed for which variable instead of the variable name.

stat_labels

A list in which is specified which label should be printed for which statistic instead of the statistic name.

diagram

Function that provides the generation of the main diagram area. Built in are: dg_vbars.

visuals

Visual parameters set with graphic_visuals().

axes

Axes parameters set with graphic_axes().

dimensions

Dimension parameters set with graphic_dimensions().

fine_tuning

Fine tuning parameters set with graphic_fine_tuning().

add_texts

Use the add_textbox() function to freely place one or multiple textboxes.

add_forms

Draws forms freely onto the canvas. Add forms with: add_line().

output

Output parameters set with graphic_output().

na.rm

FALSE by default. If TRUE removes all NA values from the variables.

print

TRUE by default. If TRUE prints the output, if FALSE doesn't print anything. Can be used if one only wants to catch the output data frame and workbook with meta information.

monitor

FALSE by default. If TRUE, outputs two charts to visualize the functions time consumption.

pct_group

If pct_group is specified in the statistics, this option is used to determine which variable of the row and column variables should add up to 100 %. Multiple variables can be specified in a vector to generate multiple group percentages.

Value

Returns a list with the data table containing the results for the graphic and the final grid::gTree graphics object.

Details

design_graphic() is actually not based on any 'SAS' function. It is envisioned to generate multiple different graphic iterations in no time. From a user perspective one shouldn't have to think about "Where is the button to set the option x to y?", "In which place do I have to put which piece of code to eventually make x happen?" or "In which order to I have to draw which things so that hopefully the right graphic prints on screen?". As a user you should only worry about "Do I use vertical or horizontal bars?", "Do the colors have enough contrast?", "Do I use a legend or direct segment labels?", "Should values be displayed or not?", "Should the graphic be static or interactive?".

The user shouldn't be a graphics engineer, but rather a graphics designer. So the function design_graphic() takes care of all the tedious work and only leaves you with the fun part.

See also

Examples

# Example data frame
my_data <- dummy_data(100, insert_na = FALSE)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 65" = 18:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Male"   = 1,
    "Female" = 2)

# Design grouped vertical bar chart
set_titles("The qol-package brings powerful concepts from 'SAS' to 'R'.")
set_footnotes("To make life easier and produce bigger",
              "and more complex outputs in less time",
              "with less code.")

my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use multiple axes grouping by just adding variables with a +
my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use multiple different segment variables within the same grouping
my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex + education",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use different stylings. With the set_graphic_options() function the styling
# can be set globally.
set_graphic_options(primary_axes_max    = 100,
                    primary_axes_suffix = " %",
                    color_theme         = "violet_fire",
                    title_font_color    = "#FF00FF")

my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Another possibility is to directly plug the styling into the corresponding
# function parameters.
my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.),
                    dimensions = graphic_dimensions(graphic_width  = 18,
                                                    graphic_height = 12,
                                                    margins        = 0.4))