Set Global Graphic Options

Modify styling options used by design_graphic(). Available parameters can be seen in graphic_visuals(), graphic_axes(), graphic_dimensions(), graphic_output() and graphic_fine_tuning().

set_graphic_options() sets the graphic options globally. These options are used by all functions involved in graphics production.

get_graphic_options() Prints out the currently set global graphic options.

reset_graphic_options() Resets global graphic options to the default parameters. This includes all options set with set_graphic_options(), set_labels(), set_titles() and set_footnotes().

Usage

set_graphic_options(..., save_file = NULL)

get_graphic_options(from_file = NULL)

reset_graphic_options()

Arguments

...

Put in any graphic option from graphic_visuals(), graphic_axes(), graphic_dimensions(), graphic_output() or graphic_fine_tuning() with the new value.

save_file

A full file path to an RDS file in which global graphic options should be stored.

from_file

A full file path to an RDS file in which global graphic options are stored.

Value

set_graphic_options(): Returns modified global graphic options.

get_graphic_options(): List of global graphic options.

reset_graphic_options(): Returns default global graphic options.

Examples

# This function can process any parameter from graphic_visuals(), graphic_axes(),
# graphic_dimensions(), graphic_output() or graphic_fine_tuning() and sets the
# option globally.
set_graphic_options(title_font_size = 12,
                    label_font_size = 10)

get_graphic_options()

reset_graphic_options()